GenAI Marketview/stock_analysis/requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
plotly>=5.14.0
yfinance>=0.2.0
wikipedia-api>=0.6.0
requests>=2.28.0
beautifulsoup4>=4.11.0
boto3>=1.26.0
python-dotenv>=1.0.0
finnhub-python>=2.4.0